Fix null expcetions

+++ app/services/update_home.rb
@@ -13,6 +13,8 @@ class UpdateHome
       delete_old_job
 
       update_home
+
+      raise ActiveRecord::Rollback if home.errors.count.positive?
     end
 
     home
@@ -23,7 +25,7 @@ class UpdateHome
   attr_reader :home, :params
 
   def delete_old_job
-    home.delayed_job.destroy
+    home.delayed_job.destroy if home.delayed_job
   end
 
   def update_home
@@ -32,5 +34,7 @@ class UpdateHome
     home.delayed_job = home.delay(cron: HOME_TOKEN_UPDATE_CRON).fetch_token
 
     @status = home.save
+
+    home.delay.fetch_token if status
   end
 end
