+++ app/services/update_triggered_task.rb
@@ -13,6 +13,8 @@ class UpdateTriggeredTask
       delete_old_job
 
       update_triggered_task
+
+      raise ActiveRecord::Rollback if triggered_task.errors.count.positive?
     end
 
     triggered_task
@@ -23,7 +25,7 @@ class UpdateTriggeredTask
   attr_reader :triggered_task, :cron, :params
 
   def delete_old_job
-    triggered_task.delayed_job.destroy
+    triggered_task.delayed_job.destroy if triggered_task.delayed_job
   end
 
   def update_triggered_task
